@@ -12,7 +12,7 @@ class Bookshelf extends Entity implements HasCoverImage
 
     protected $fillable = ['name', 'description', 'image_id'];
 
-    protected $hidden = ['restricted'];
+    protected $hidden = ['restricted', 'image_id'];
 
     /**
      * Get the books in this shelf.
