@@ -12,9 +12,9 @@
         <input type="hidden" value="{{ $order }}" name="order">
 
         <div class="list-sort">
-            <div class="list-sort-type dropdown-container" dropdown>
-                <div dropdown-toggle aria-haspopup="true" aria-expanded="false" aria-label="{{ trans('common.sort_options') }}" tabindex="0">{{ $options[$selectedSort] }}</div>
-                <ul class="dropdown-menu">
+            <div component="dropdown" class="list-sort-type dropdown-container">
+                <div refs="dropdown@toggle" aria-haspopup="true" aria-expanded="false" aria-label="{{ trans('common.sort_options') }}" tabindex="0">{{ $options[$selectedSort] }}</div>
+                <ul refs="dropdown@menu" class="dropdown-menu">
                     @foreach($options as $key => $label)
                         <li @if($key === $selectedSort) class="active" @endif><a href="#" data-sort-value="{{$key}}">{{ $label }}</a></li>
                     @endforeach
